backend/deepfake-model/requirements.txt
torch
transformers
Pillow
opencv-python
dlib
torchvision 
opencv-python 
face_recognition 
timm